requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
djangorestframework-simplejwt
coverage
requests
celery
redis
django-celery-beat
django-cors-headers
drf-yasg
flake8
